requirements.txt
# Core numerical & imaging
numpy>=1.26
pillow>=9.5
opencv-python>=4.8

# OCR (requires external Tesseract executable installed on system)
# On Windows, install Tesseract from: https://github.com/tesseract-ocr/tesseract
# The code auto-detects it or use env var TESSERACT_CMD to point to the exe
pytesseract>=0.3.10

# 3D geometry & rendering
trimesh>=4.0
pyrender>=0.1.45
PyOpenGL>=3.1.7
pyglet>=2.0

# HuggingFace VLM pipeline
transformers>=4.41
huggingface-hub>=0.23
safetensors>=0.4.3
# Choose the appropriate Torch build for your system (CPU/CUDA). Example:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121  (CUDA 12.1)
#   pip install torch  (CPU-only)
torch
